Text Mining engine using TF*IDF technique
Dear my friend, 
Hope all is well, 
Currently, you are viewing one of my recent project targeting the big data analysis and text mining in python. The attached application is a text mining engine.  It is capable to analysis the given text and extracts the meaningful and unique keywords through the submitted text. The model is capable of analysis thousands of thousand records in a minute.  Hope you enjoy of using this smart app. 
. 
How to run: Python start.py sample.txt,
How to Exit: the process would be close by typing (end-of-file: eof) “ eof”.
•	The attached files are a sample database (word.db), sample text (sample.txt), and my codes (start.py).
•	The sample file contains a set of 5000 recodes.